+++ components/audio_player/index.ts
@@ -1,2 +1,0 @@
-export { AudioContextProvider, AudioContext } from "./audio_context";
-export { AudioCover } from "./audio_cover";