+++ components/draggable_audio/index.ts
@@ -1,1 +1,0 @@
-export { DraggableAudio } from "./draggable_audio";