+++ components/draggable_embed/index.ts
@@ -1,1 +1,0 @@
-export { DraggableEmbed } from "./draggable_embed";